move to standard price
